SimpleTrainConfig is a Marin default.
