+++ experiments/dolma/exp442_dolma.py
@@ -8,13 +8,12 @@ https://github.com/marin-community/marin/issues/442
 
 from fray.cluster import ResourceConfig
 from levanter.models.llama import LlamaConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 from marin.processing.tokenize.data_configs import lm_mixture_data_config
 
 from experiments.llama import llama_1_4b, llama_1_4b_train_config
 from experiments.pretraining_datasets import DOLMA_OLMO_MIXTURE_WEIGHTS, tokenize_dolma
-from experiments.simple_train_config import SimpleTrainConfig
 
 EXPERIMENT_TAG = ["442_dolma"]
