+++ experiments/ferries/daily.py
@@ -20,11 +20,10 @@ import os
 from dataclasses import replace
 
 from fray.cluster import ResourceConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 
 from experiments.llama import compute_num_parameters, llama3_tokenizer_vocab_size, llama_150m
-from experiments.simple_train_config import SimpleTrainConfig
 from experiments.tootsie.exp1295_32b import nemotron_mix
 
 # ---------------------------
