+++ experiments/llama.py
@@ -10,8 +10,7 @@ from levanter.data.text import ChatLmDatasetFormat
 from levanter.layers.rotary import Llama3RotaryEmbeddingsConfig
 from levanter.models.llama import LlamaConfig
 from levanter.utils.activation import ActivationFunctionEnum
-
-from experiments.simple_train_config import SimpleTrainConfig
+from marin.defaults import SimpleTrainConfig
 
 llama3_tokenizer = "meta-llama/Meta-Llama-3.1-8B"
 llama3_tokenizer_vocab_size = 128_256
