+++ experiments/plantcad/exp1729_plantcad_train.py
@@ -11,11 +11,10 @@ import jax
 from fray.cluster import ResourceConfig
 from levanter.data.text import TextLmDatasetFormat
 from levanter.models.llama import LlamaConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 from marin.execution.types import versioned
 
-from experiments.simple_train_config import SimpleTrainConfig
 from experiments.tokenization import default_tokenize
 
 logger = logging.getLogger(__name__)
