+++ experiments/scaling_law_sweeps/c_adamc.py
@@ -32,14 +32,13 @@ from levanter.layers.rotary import Llama3RotaryEmbeddingsConfig
 from levanter.models.llama import LlamaConfig
 from levanter.models.qwen import Qwen3Config
 from levanter.optim.cautious import CautiousConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.types import ExecutorStep, InputName
 from marin.processing.tokenize import get_vocab_size_for_tokenizer
 from marin.scaling_laws import CandidateConfig, pick_v5p_type
 
 from experiments.evals.evals import default_eval
 from experiments.evals.task_configs import EvalTaskConfig
-from experiments.simple_train_config import SimpleTrainConfig
 
 # --- Constants ---
 SEQ_LEN: int = 4096
