+++ experiments/scaling_law_sweeps/completed_adamh.py
@@ -41,7 +41,7 @@ from levanter.layers.rotary import Llama3RotaryEmbeddingsConfig
 from levanter.models.llama import LlamaConfig
 from levanter.models.qwen import Qwen3Config
 from levanter.optim import AdamHConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.types import ExecutorStep, InputName
 from marin.processing.tokenize import get_vocab_size_for_tokenizer
 from marin.scaling_laws import CandidateConfig, pick_v4_type
@@ -49,7 +49,6 @@ from marin.scaling_laws.tpu_utils import V4_SPEC
 
 from experiments.evals.evals import default_eval
 from experiments.evals.task_configs import EvalTaskConfig
-from experiments.simple_train_config import SimpleTrainConfig
 
 # --- Constants ---
 SEQ_LEN: int = 4096
