+++ experiments/tootsie/exp600_tootsie.py
@@ -22,7 +22,7 @@ import dataclasses
 
 from fray.cluster import ResourceConfig
 from levanter.schedule import ScheduleStep
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 from marin.processing.tokenize.data_configs import lm_varying_mixture_data_config
 
@@ -43,7 +43,6 @@ from experiments.pretraining_datasets.dclm import (
     dclm_components_llama3,
     dclm_mixture_config_llama3,
 )
-from experiments.simple_train_config import SimpleTrainConfig
 
 # Phases/Runs in this file:
 # 1. Kestrel: WSD-S on DCLM+Starcode+Proofpile on 2x v5litepod-256 (from scratch)
