+++ experiments/tootsie/exp750_tootsie70b.py
@@ -20,12 +20,11 @@ import dataclasses
 
 from fray.cluster import ResourceConfig
 from levanter.schedule import ScheduleStep
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 
 from experiments.llama import llama_70b
 from experiments.pretraining_datasets.dclm import dclm_mixture_config_llama3
-from experiments.simple_train_config import SimpleTrainConfig
 from experiments.tootsie.exp859_big_tootsies import dclm_mixture_config_llama3_zoned
 
 llama_70b_train_config_mk6 = SimpleTrainConfig(
