+++ experiments/tutorials/exp1077_reproduce_dclm_1b1x.py
@@ -18,10 +18,9 @@ Example usage:
 
 from fray.cluster import ResourceConfig
 from levanter.models.llama import LlamaConfig
-from marin.defaults import default_train
+from marin.defaults import SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 
-from experiments.defaults import SimpleTrainConfig
 from experiments.pretraining_datasets.dclm import dclm_mixture_config_llama3
 
 # Define the LlamaConfig for a 1.4B parameter model
